+++ remotedb/batch.go
@@ -15,21 +15,23 @@ type batch struct {
 var _ db.Batch = (*batch)(nil)
 
 // Set implements Batch.
-func (b *batch) Set(key, value []byte) {
+func (b *batch) Set(key, value []byte) error {
 	op := &protodb.Operation{
 		Entity: &protodb.Entity{Key: key, Value: value},
 		Type:   protodb.Operation_SET,
 	}
 	b.ops = append(b.ops, op)
+	return nil
 }
 
 // Delete implements Batch.
-func (b *batch) Delete(key []byte) {
+func (b *batch) Delete(key []byte) error {
 	op := &protodb.Operation{
 		Entity: &protodb.Entity{Key: key},
 		Type:   protodb.Operation_DELETE,
 	}
 	b.ops = append(b.ops, op)
+	return nil
 }
 
 // Write implements Batch.
@@ -49,6 +51,7 @@ func (b *batch) WriteSync() error {
 }
 
 // Close implements Batch.
-func (b *batch) Close() {
+func (b *batch) Close() error {
 	b.ops = nil
+	return nil
 }