+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### Breaking Changes
 
+- `Batch` interface methods now return errors
+
+- Removed the `SetDeleter` interface
+
 - [memdb] [\#56](https://github.com/tendermint/tm-db/pull/56) Removed some exported methods that were mainly meant for internal use: `Mutex()`, `SetNoLock()`, `SetNoLockSync()`, `DeleteNoLock()`, and `DeleteNoLockSync()`
 
 ### Improvements